Exlcude CNAME, optionally disable delete flag
Some people complained that the CNAME file should not be ignored, so I
added a flag to optionally enable the sync of that GitHub specific
config.

Beside that some people also requested to optionally disable the delete
flag of the rsync command executed by this plugin, there is a flag for
that now.