Adds PLUGIN_COMMIT_APPEND_TO_MESSAGE which appends to the commit message
I've got a pages build process that is subject to some automated systems
and because we have the string "password:" in our documentation, we have
to add something to every commit that pushes to any branch in order to
disable the scan. This enables us to hardcode the disabling text in our
CI configuration.

Also, adds a test because I don't feel confident enough in my Go skills
to do anything without tests.