Reboot the plugin to use newest conventions (#34)
* Reboot the plugin to use newest conventions

Remakes the gh-pages plugin to use the latest conventions for Drone plugins. Drops use of `urfave/cli` for a pure cli app. Also supports cards.

* Allow commit message to be specified

When `message` is specified it can be a URI, a file path, or a string. If its a URI with a http(s) scheme a request will be made and the response body will be used as the commit. If its a URI with a file scheme or the path to a file that file is loaded and its contents used as the commit message. This allows for a commit message to be generated in a previous step. If its just a string that is used as the message. If its an empty string use the same commit message on the pages branch.

* Update to go 1.20